Introduce a CLI script to generate PHPUnit config
This replaces the existing dynamic suites, that are no longer supported
in PHPUnit 10 and are also not supported by paratest.

With this patch, there are two steps to running PHPUnit tests: first, we
run the `generatePHPUnitConfig` script, which reads LocalSettings.php
and gets a list of installed extensions/skins, as well as a list of test
paths (which is `$ext/tests/phpunit` for every installed ext/skin, plus
any modifications done by the UnitTestsList hook). It then writes this
list to a local-only phpunit.xml file, based on the existing
phpunit.xml.dist file (which, on its own, cannot be used to run
extension tests). Then, we run PHPUnit normally using the newly
generated config file.

Add a structure test to verify that the local config exists and is
up-to-date.

Parser tests could also be migrated to this approach, but that's a
bigger change and is left for another day.

Bug: T345481
Change-Id: I44ef94cdb95aa5d6be185c3e37a7958f0fbc3b64