Make StatusValue stable to extend
Custom Status or StatusValue subclasses can be useful to make the
contained value easier to work with – see e.g. PageUpdateStatus with its
custom factory newEmpty() and the accessors getNewRevision(),
wasPageCreated() etc. Let’s bless this practice by marking StatusValue
as stable to extend; Status is left untouched, because we generally want
to encourage use of StatusValue, not Status, these days. (Note that both
classes are already extended in some extensions, namely MediaModeration,
ReportIncident and SecurePoll.) Several protected fields by the class
which are not meant for other subclasses are marked @internal.

Bug: T358492
Change-Id: I6690fbbb13624ce650d45ab2a872ba38f1dd90a2