Make chunked upload delete chunks in deferred update
I believe deleting chunks of large uploads may have been taking
a fair bit of time, potentially triggering a timeout causing the
whole job to fail (Which is unfortunate as they will just be
deleted by a cron job later). Putting them in a deferred job
after transactions commit should ensure that the upload will
still go through.

Additionally, this ties it to AutoCommitUpdate which should cause
the deletes to be cancelled if something causes the transaction
to abort, which is a step towards making the Assemble job be retriable
on failure.

Bug: T200820
Change-Id: I49b8eb23adcd8960783f4f90707faa21760ce078