phpunit: Do not crash when paratest options are passed in
Paratest takes additional options that are not recognized by PHPUnit.
While paratest does unset them when invoking PHPUnit, we were previously
passing them through in the bootstrap file, which results in an uncaught
exception.

So, make a list of a few known paratest options and make PHPUnit ignore
them.

This is obviously doomed to fail miserably as soon as someone tries to
pass another paratest option that is ot listed here, but since we aren't
officially supporting paratest at the moment, it should be fine for the
time being.

Bug: T50217
Change-Id: I4c197676d9410adebc47baf50c699ec7cb411ed8