maintenance: show broken replicas in getLagTimes
When replication is broken to a host, its lag is returned as `false`
which was converted to an integer and thus yielding 0. That hides the
fact the replication is broken, specially when all other replicating
hosts have sub second lag.

Give a meaningful message when replication is broken, same as we did for
lag.php in Ibc88d2b86384a68b3cf6fe0d9739144247534821.

Bug: T358484
Change-Id: I321516cf054f23940e0f0944832b57a387434721