No need to crash the Parser on old revs with broken UTF-8
The argument is the same as in 6f32dc8a8db. Full quote:

"Garbage in, garbage out. When the wikitext is broken, it's still
helpful if the user can see the broken wikitext. Even if it's not
fully parsed. It's not the job of this class to fix broken UTF-8.
The worst thing that can happen is that the wikitext contains some
unparsed magic words. However, this is really only relevant for
very old revisions (20 years old, see T321234). It's very normal
that old revisions can't be 100% parsed any more, most notably
because of deleted templates. This here is not much different."

While we might care about broken UTF-8, it's not the job of the
code touched in this patch but needs to be done somewhere else.
I believe there is no added value in making the code crash in
these specific spots.

Bug: T321234
Change-Id: I0be8d7bae8dc1170ace1adf157f48a0145e15938