CodexModule: Explicitly disable URL loading
FileModule::supportsURLLoading() looks like it should return false for
CodexModules since they have packageFiles, but it runs before we can add
things to $this->packageFiles (at least for modules without packageFiles
of their own, like @wikimedia/codex-search). We can't call setupCodex()
and then call the parent method like we do for other overridden methods,
because we don't get a Context object, so just always disable URL
loading.

Not doing this led to a bug where the output of the
@wikimedia/codex-search module was empty when ?debug=1 was passed,
because we erroneously marked that module as supporting URL loading when
in reality it contained package files.

Bug: T358903
Change-Id: I7b87f78ed901b79236c8280a08bb73ac813926c7