Use transactional time limit on api Upload (except checkstatus)
Uploads can be a long operation that we would prefer not to get
cut off. The API upload module supports a variety of upload types.
This calls useTransactionalTimeLimit() for all of them (except
checkstatus which is not a write operations) however it is more
applicable to some methods than other methods such as "async" where
the real work happens in the job queue.

Change-Id: I6795f6d38693306f22091afa77123f2f0721ef5b