parser: Remove LoggerSpi override from ParserOutputAccessTest
The presence of this mock means that, valuable debug messages are
missing from PHPUnit output. Considering how high-level that is,
that is quite a bit.

Follows-up a longer-than-needed debug session from I5413825a0
and Ie532c17e5b. Most of the time was actually not spent adding this
back in, but trying to figure out how it was that the messages
were missing.

Bug: T358901
Change-Id: I41fa8dce1ea37ea69632917bcb591b8327d35e7d