resourceloader: Update mw.loader tests to always use full URL
Some requests used $wgServer and some did not. Make this consistent,
and also make it more robust against unusual $wgScriptPath configurations
such as empty string (doc root install) and full URL (different domain),
by letting the native URL class do the relative expansion instead of
unconditionally concatenating the two.

This is prep for removing the need for a Karma proxy from "/scriptpath"
to "http://server/scriptpath". This was the last remaining test in core
that relied on network requests, and did so directly to wgScriptPath.

Change-Id: Id38f811dc945afa32370089adc0f8111444df818