Make sure MovePage::isValidFileMove matches UploadBase::getTitle
MovePage::isValidFileMove was not checking that filenames are at
most 240 bytes. This was causing metadata for old versions of
file to be cut off.

Also add checks for non-ascii filenames on systems not supporting
unicode. To be extra careful also add an explicit check for files
that are only an extension although its likely such moves would
have failed for other reasons.

Bug: T359294
Change-Id: I9e75a3e4007b76d66f1305e9107dadbbbfb4d5cb