build: Simplify --qunit-watch logic
* Remove limitation to `.js` files, so that edits to Resources.php
  and any .json files are considered as well.

* Add watching of the skin.json and extension.json file itself.

* Allow `--qunit-watch` to work without `--qunit-component`,
  defaulting to watching MediaWiki core.

I would have liked to widen this to simply watch `extensions/**` and
`skins/**` by default (and when component != MediaWiki core), because
that way it will also re-run automatically after changes in an
extension dependency (e.g. from core to an extension, or
from a dependency like EventLogging to NavigationTiming). However,
Karma uses the popular "chokadir" which ironically chokes and runs out
of available operating system file handles in an attempt to crawl and
watch every file recursively. This unlike qunit-cli which uses
node-watch and the Node.js 12+ native file watcher API. Perhaps for
something to revisit in the future.

Change-Id: I2bb7f0df510f84872a2381bdebe3f2b59cc80705