ResourceLoader: Remove mw.track stub to simplify mw.trackError
Instead of merely saying we shouldn't use mw.track internally,
actually make it undefined until mediawiki.base loads.

Internally we can push to mw.trackQueue, the format is simple enough
and a copy of this will exist in the startup module either way.

The indirection and complexity to save 1 line of code is not worth
the cost. The real savings of moving mw.track from mediawiki.js
to mediawiki.base.js, came from making mediawiki.js independent
of jQuery and $.Callbacks, and thus allowing jquery and mediawiki.base
to load in parallel (T192623). The line of code to define mw.track
is not what we save.

While at it, remove an unused parameter to the internal logError()
function.

Change-Id: Ie5b8ae4bb70beb8c706f73d01366870c13503a27