Don't strip non-newline whitespace from left side of [[Category]] links
This follows up on a long series of tweaks to whitespace handling around
[[Category]] links (T2087, T87753, T174639) which aimed to simplify and
make intelligible the whitespace handling around category links without
allowing categories to break lists or paragraphs in which they are found.

Removing newlines but not other whitespace on the left-hand side of
category links should preserve the valuable features of T2087 et al
while still ensuring that the following all render equivalently:

  ABC [[Category:Foo]]DEF
  ABC[[Category:Foo]] DEF
  ABC [[Category:Foo]] DEF

Added parser test to document the new behavior; it's worth noting
that although there were plenty of tests documenting the expected
interaction of category links and newlines, there were previously
no tests covering the interaction of non-newline whitespace and
category links; the one test which needed to be altered added
non-semantic whitespace (ie, extra whitespace to the test output
which did not affect the way the HTML would display).

This patch brings the legacy parser into parity which Parsoid parsing
of category links.

Bug: T359886
Change-Id: I5e87b33a956e296cdaf671fa99c9555944b73479