HTMLFormField: Add Codex version of Select-related fields
This patch enables use of a Codex component for the following
HTMLFormFields:
- HTMLExpiryField
- HTMLSelectAndOtherField
- HTMLSelectField
- HTMLSelectOrOtherField

It also adds a new public static method to HTMLTextField that
takes in attributes and returns the HTML for a Codex text input.

Note that the Codex CSS-only Select does not yet support an
error state, so this has not been implemented here.

Bug: T359015
Change-Id: I6857813b9dbe09ae68d6cafdc87351d01626266a