filebackend: Fix MemoryFileBackend move result
MemoryFileBackend, a class which is only used for testing, does not
correctly respect the ignoreMissingSource option to its move operation.
The copy succeeds due to the forwarded ignoreMissingSource option, so it
proceeds with the source deletion, which fails.

This was implemented in the base class method FileBackendStore
::doMoveInternal, but every other subclass was overriding it.

So, make it abstract, by analogy with doCopyInternal and
doDeleteInternal. Add an implementation specific to MemoryFileBackend
which handles ignoreMissingSource correctly.

Change-Id: Ib5733046d741b962c65db42b5364705b073c4b06