Edit Recovery: always store and don't restore on POST
Always save the form data to indexedDB even if it's the same as
the 'original' data (i.e. the form data as it is when the page
is lodaed). This means that for non-live previews, switching from
VisualEditor, etc., the POSTed state of the form will be saved.

In order to make it possble to do this, remove the delete-unchanged
logic, and rely on detecting the unchanged state when we restore
the data. This does mean that some redundant data will be stored,
but it makes it more robust in other situations and so this is
worth it.

Bug: T347864
Change-Id: I37bb40f9f4c8c87148290f9deaa25f8b6a3b7c1c