ResourceLoader: Refactor makeTitleInfo() in WikiModuleTest
Use TitleValue instead of Title::newFromText to reduce complexity of
this test.

Change `Project:Example/foo.js` example to NS_USER to be more
realistic, given that (afaik) there can't be a WikiModule with a
page in NS_PROJECT. It is true that people sometimes import project
pages containing JS via importScript over action=raw, but that
doesn't involve ResourceLoader.

Change-Id: Id2df9a0f17880e397b02d4c855d9e0287a8949ee