REST: Accept POST with empty body and no content-type
All POST requests must have a body, even if the body's is
empty, with content-length set to 0. When the body is empty, it should
not be necessary to specify a content-type either.

A typical use case would be some kind of "touch" semantics that would
update a resource's state without needing any additional information
from the client.

Change-Id: I7d20bb6a8c484b9b4725f056342ff8d015a171ac