block: In UsersPager use a subquery instead of HAVING()
To filter or detect deleted users, use the subquery from HideUserUtils
like everything else, instead of a weird special solution just for
UsersPager, which turns out to be slow.

The solution here is very similar to ApiQueryBlockInfoTrait.

PostgreSQL makes aggregation a bit awkward. See the comment for the
full story.

Bug: T360160
Change-Id: Ifd365ca7fceae5edab7ebbff33fb80485e09ddfd