i18n: make chn "mainpage" a valid title
The incubator wiki has some backend errors for the chn wiki since the
main page is an invalid title. The reason is the "mainpage-description"
got amended to add description which includes a newline and that got
mistakengly copy pasted to "mainpage". The later is used to forge a
Title and the newline character make it an invalid Title.

The faulty localization comes from Feb 22nd 2024 with:
cdabac07fbd09cb254da024da9a4f72268913180

Bug: T360066
Change-Id: I59118b9892d11ad7b26d5ebdeccab82d98b28fc7