HTMLForm: Use rows=10 instead of rows=25 for Codex multiselect fields
The Codex implementation of HTMLNamespacesMultiselectField and
HTMLTitlesMultiselectField falls back to a simple `<textarea>`, which is
also what the OOUI implementation does if JS is not available. However,
OOUI sets rows=10, which results in a much more reasonably sized
textarea then HTMLTextAreaField's default of rows=25.

Pass rows=10 to HTMLTextAreaField, but allow the field params to
override this.

Change-Id: I1e9a1ab6b4a71fb247bfe7861a49bb6cfeb61f54