SpecialBlock: Use option descriptions instead of label hack
Use the new support for per-option descriptions in CodexHTMLForm to
replace the hacky solution that was previously used. This is only used
when the UseCodexSpecialBlock feature flag is enabled, because only
CodexHTMLForm supports per-option descriptions; if the feature flag is
disabled, then we're still using OOUIHTMLForm, and we need to use the
old label hack.

Bug: T359012
Change-Id: I65caf6096553506f9e262a8e56833935701143d8