rdbms: In LoadMonitor, use a loop of getWithSetCallback
LoadMonitor has a server cache (APCu) layer including locks for
stampede protection, in front of WANObjectCache with getMulti().
Apparently getMulti() was used instead of getWithSetCallback() to
reduce latency, but this seems poorly justified for such a
well-protected cache miss.

Simulation results showed that all server caches expire at once, and
there are as many simultaneous metric probes as there are MW hosts. The
randomness in isStateFresh() was not effective.

So, instead of getMulti(), use a loop of getWithSetCallback() with the
lockTSE option to avoid a stampede. Remove the random refresh from
isStateFresh().

Bug: T314020
Change-Id: Icd51680b557834d405a9d5297c1b189a0d1a44a8