Update eslint-plugin-vue to 9.23.0
In theory we shouldn't need to do this, because eslint-config-wikimedia
already provides eslint-plugin-vue. However, eslint-plugin-mediawiki
also pulls in eslint-plugin-vue but at an older version, and this causes
nonsensical vue/dot-location errors to be raised at the end of the
`<script>` block in each .vue file.

This is a temporary fix to make it possible to introduce .vue files in
MW core without having to disable the vue/dot-location at the end of
each file. Once this is fixed permanently upstream (see
https://github.com/wikimedia/eslint-plugin-mediawiki/pull/95), this
change can be undone.

Change-Id: Ibd616750f046a3a66b5d7f2665696ed9175a4090