CodexHTMLForm: Codexify buttons at the bottom of the form
Introduces a new public method CodexHTMLForm::getButtons
that provides a way to add Codex buttons to the bottom
of HTMLForm. This includes the submit, cancel, and
reset buttons. These buttons are wrapped in a <span>
with the class name 'mw-htmlform-submit-buttons'.

Note:
- The reset button is not included in this patch because
it has been unused and we plan to remove it (T361032).
- Temporary fix to resolve PHP warnings when 'fieldname',
'name', and 'id' are not set. We plan to follow-up by
changing HTMLButtonField to use a static method instead.
This will also resolve cases when tooltip is not set.

Bug: T359017
Change-Id: If064b1705d2e3c8f075e5f8a2194e9fffeae8675