Special:BlockList: apply simpler conditions when listing user blocks
When both IP blocks and range blocks are hidden, only user blocks are
left, but the DB is confused by this and puts block_target first in the
join, sorting by bt_user, copying 800k rows to a temporary table. With
a more explicit condition on bt_user, the incorrect query plan is
avoided.

Bug: T360864
Change-Id: I8f863124cfb46ffe46a30b38b9a538c6a7aa3d4d