rdbms: Set up section-wide circuit breaking
This makes sure in case one section is overloaded, appserver workers
stop sending more requests to that section and simply fatal (fast) which
prevents appserver resource exhuastion.

Based on our measurements, 400 is pretty good number to trigger this. If
anything, it's a bit too high. Will see if that's the case in the next
incident.

Bug: T360930
Change-Id: I2f3a2abe8f0f76aea8a234ffaddf0c9d8040e510