HTMLForm: Make raw InfoFields work in Codex
HTMLInfoField overrides getDiv(), getTableRow() and getOOUI() to output
the raw value if 'rawrow' is set. Override getCodex() too, so that
'rawrow' also works correctly in CodexHTMLForm.

This caused a bug where going to
Special:UserLogin?warning=mobile-frontend-watchlist-purpose would show a
message box that was incorrectly wrapped in a cdx-field div. That seems
to have been harmless, but it's incorrect behavior and could cause
styling bugs in the future.

Change-Id: I1fd4e2e53f9e03561fe300b6924e44cf222db3c9