SpecialCreateAccount: Prevent JS from un-bolding the email label
Adapt the JS to the new DOM structure created by the conversion of the
login / account creation form to Codex. Because it was assuming the old
DOM structure, it unwrapped a `<span>` tag and caused the label for the
email field to no longer be bolded.

In the longer term, we should add support for optional flags to
CodexHTMLForm, and once that has been done we should be able to just
show/hide the optional flag separately, which will allow us to simplify
this code and also stop using two different messages for the optional
and non-optional versions of the email label.

Change-Id: I68fe26be9a17f3fa30658f2456aa43fe7d88e018