checkComposerLockUpToDate.php: Fix MessageCache crash during update.php
When running update.php after a git-pull in core or an extension,
the localisation cache is considered out of date and thus needs a
refresh, which includes contacting the database, which is problematic
when we're still running update.php to update that same database.

```
Uncaught LogicException: Process cache for 'en' should be set by now
```

This is similar to the CI failure that the "isQuiet" mode addressed
for running PHPUnit, except it merely hid the problem because the same
can happen when we're not in a quiet mode.

Change-Id: I33bf81087c282e0bcb5a0fc70ae31589cf72e7bb