composer: Remove localization from unlocalizable LockFileChecker
This is internal class is exclusively used by LockFileChecker,
which in turn is used by maintenance/checkComposerLockUpToDate.php,
which in turn is used by PHPUnit and maintenance/update.php.

Both of these cases have to use workarounds to disable localisation,
because they run on the CLI (no request context or user language),
and without access to MediaWikiServices, Database, or LocalSettings.

Follows-up:

* T330228: I2caddfb8f2 (5ecbb62f7f): Add composer check to PHPUnit.
* T333412: Skip composer check for mediawiki-vendor Jenkins jobs.
* If6110eeaec (397e00444a): Fix invisible pre-fatal errors messages.
* T283389: I0d56f23560 (20e77793b8): Refactor and add localization.
* I995a1cb01a (d5d585b9b2): Fix "premature access" during PHPUnit.
* I33bf81087c2: Fix MessageCache crash during update.php.

The localization was added for T283389, but ended up not using it,
since all it does is check isGood and then render a custom interface
message localised as part of the Installer. Hence, remove this to
categorically remove and avoid these problems.

This change effectively reverts I995a1cb01a (d5d585b9b2) and thus
restores the more detailed and useful error details when failing
in the context of a PHPUnit command.

Change-Id: Ia30d39d7e064e9ff0d8397d25918d0bb2ea10914