HTMLRadioField: Fix escaping of options with Codex
`$this->mOptionsLabelsNotFromMessage` is set when calling
`$this->getOptions()`, meaning the previous code wasn't actually seeing
its expected value. Compare with the `getInputOOUI` implementation,
where the ternary is inside the foreach loop.

So, do the same thing for Codex and choose the function to use inside
the loop. Also use the full method name in both cases instead of
indirect calls, as it makes the code easier to analyze statically (also,
the call itself is pretty simple).

Follow-up to I73b2e95cff.

Change-Id: Ie6b626fa4f9822f76bf202c5ed66ec16c17dc422