ResourceLoader: Inline logError and simplify mw.trackError further
Remove handler from Special:JavaScriptTest. These are rare to happen
during development, other than when developing RL itself, for which
we have better tests already (and they show up in the console either
way, we don't have to rely on QUnit marking them as failures).

This handler could be made to work, but then we'd need a way to
opt-out during meta tests for mw.trackError itself, which doesn't
seem justified in terms of complexity.

Change-Id: I7be11ebe942d35a437b14245f7d8a4f36f3fe7ce