filebackend: simplify "headless" parameter and STREAM_HEADLESS flag
In FileBackend::streamFile(), do not send "headers" in "headless" mode.

In HTTPFileStreamer::stream(), do not treat $headers as an exception.
Just use $headerFunc like with all the other headers.

Nothing relies on the old behavior, and the only things using "headless" are
integration tests anyway. The same holds true for STREAM_HEADLESS.

Change-Id: I75faf9195a713d6a8a8f1273cb9e747980f4d57f