Add additional PathMatcher test cases
In T359652 we are contemplating disallowing optional path
parameters in the REST API. Related investigation shows we are
lacking test cases for some current edge case path matching
behaviors. Add tests for these, to make the current behavior
explicit. If we decide to disallow optional path parameters,
these new tests will need to be revised. This is good. A change
in behavior should result in failing tests, and without these
new test cases, it would not.

Bug: T359652
Change-Id: Ibc0880ba0f6ea199a91fc7fc81577519989d32df