GitInfo: Fix Phan warning on Windows
On my Windows machine, I was getting the following Phan warning:

includes\utils\GitInfo.php:159 UnusedPluginSuppression
  Plugin BuiltinSuppressionPlugin suppresses issue
  PhanPluginDuplicateArrayKey on this line but this
  suppression is unused or suppressed elsewhere

Indeed, this isn't a PhanPluginDuplicateArrayKey issue on Windows,
since DIRECTORY_SEPARATOR is `\`; but it is a legitimate warning
on other systems where DIRECTORY_SEPARATOR is `/`.

Refactor the code slightly so as to not create an array with
potentially duplicate keys.

Change-Id: I4be38aa9fefde26d58139d74a85359961229baec