CodexHTMLForm: Use a static method in getButtons()
Create a static method that returns the HTML for a Codex button.
CodexHTMLForm::getButtons() and HTMLButtonField::getInputCodex()
calls the static method and passes in a list of attributes,
flags, and the button label needed to build the Codex button.

This patch changes the way the Codex submit button is created. In a
previous patch, the submit button was an instance of HTMLSubmitField.

Bug: T361406
Change-Id: I046de61d4808f034335254e9a9c853f133d66a9b