Edit Recovery: return to deleting stored data if unchanged
This partially reverts I37bb40f9f4c8c87148290f9deaa25f8b6a3b7c1c
to reinstate the deletion of the indexedDB data if the form
fields are unchanged. However, it adds a check to the condition
for this to never delete if it's a POST request (such as for a
preview).

This also removes the `mode` field from the stored data as its
not related to page data and so was causing false positives in
`isSameAsOriginal()` checks.

Bug: T347864
Change-Id: Iba6d7172ab9117ac17b645f3e528ada6b01d0745